requirements.txt
pandas
streamlit
numpy
matplotlib
openpyxl
pydrive2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-secret-manager
